pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "aloha-augment"
version = "0.1.0"
description = "Augmentation pipeline for the LeRobot ALOHA dataset"
requires-python = ">=3.10"
dependencies = [
    "lerobot",
    "opencv-python",
    "numpy",
    "scipy",
    "rich",
    "fire",
    "torch",
    "huggingface_hub",
    "pandas",
    "pyarrow",
]

[project.optional-dependencies]
clip = ["transformers", "Pillow"]
dev  = ["pytest"]

[project.scripts]
aloha-augment = "aloha_augment.pipeline:main"

[tool.hatch.build.targets.wheel]
packages = ["src/aloha_augment"]

[tool.pytest.ini_options]
testpaths = ["tests"]
